Remove the `.deprecated` redirection feature
Summary
=======

The `{.deprecated: [old, new].}` feature, originally introduced for
providing information to `nimfix`, can be implemented as a library
solution and doesn't need compiler support.

Since `skAlias` was only used for these redirections, the symbol kind is
also removed, which simplifies lookup a bit, as alias symbols no longer
need to be considered there.

Details
======

The only user of the redirection feature was the `mysql` module; all
usages there are now removed, meaning that code still using the
deprecated enum names needs to be adjusted.

- remove the now unused `adSemBadDeprecatedArgs` and
  `rsemBadDeprecatedArgs` diagnostics
- update the `rsemBadDeprecatedArg` message