front: rework `--eval`
Summary
=======

The `--eval` option was previously responsible for two things:
- if no command was specified yet, changing the command to `e`
- overriding the content of the main module with the given argument

Its implementation was scattered and the behaviour not very intuitive,
so the option is simplified:
- it no longer accepts an operand, but instead uses the second command-
  line argument as the main modules content
- it doesn't set the active main command anymore

In addition, the option is renamed to `fromcmd` in order to better match
its new behaviour. Example usage: `nim e --fromcmd "echo 1"`

One thing that is now no longer possible is overriding the contents of
the main module while still using the provided project path.

Details
=======

- instead of multiple booleans, the source mode for the main module is
  now provided via an enum
- remove the now unused state previously required by the `--eval`
  implementation
- perform project name and path initialization in a single place
  (`processCmdLineAndProjectPath`)
- adjust the `truner.nim` test, which made use of `--eval`

NimScript execution
-------------------

When executing a NimScript file by either using the `e` command or only
providing the path of a NimScript file, running the script now happens
when during main command processing, instead of during configuration
loading.

Misc
----

Remove the `IdGenerator` parameter from `runNimScript`. The
`IdGenerator` of the `ModuleGraph` used for executing the NimScript is
passed to the VM context setup instead, which is a step towards
sandboxing/isolating NimScript execution.