sem: fix the compile-time-only detection
Summary
=======

The `inStaticContext` state that is, among other things, used to track
whether the currently analysed code is only available at compile-time
was part of the shared context (`PContext`), which is incorrect, as it
means that the bodies of generic procedures instantiated in, for
example, a `static` block are treated as being inside a compile-time-
only context (which they are not).

Instead, the `inStaticContext` state is now part of `TProcCon`, and
thus local to the analysis of a routine.

Details
=======

- always mark macros as being available compile-time only. This makes
  checking for compile-time-only routines easier, as macros no longer
  need to be tested for separately