sempass2: fix illegal capture detection
Summary
=======

The illegal capture detection treated all routines enclosed by a
`.compileTime` routine as also being being compile-time-only. This to
illegal code (according to the current language rules) not being
rejected.

The rules are updated and the detection adjusted to work according to
them; see the changes to the related tests for details.

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>