vm: fix multiple `raise` and `finally` issues
Summary
=======

Fix multiple issues with `raise`-related stack unwinding and resuming
after `finally` clauses. In addition, prepare for further `finally`
fixes.

Details
=======

When leaving a `finally` section, resuming the stack unwinding started
by a `raise` is now handled directly instead of via the indirect back-
and-forth mechanism used previously.

As a consequence, cactus-like stack shapes are now no longer possible,
which simplifies all logic related to traversing the stack and also
makes future VM changes that expect a "real" stack possible. The top-of-
stack also no longer has to be tracked, but is now known to always be
the last element int the stack-frame sequence.

In addition, the saved program counter (which is now only used during
frame cleanup) is stored separately for each stack-frame. This is
necessary for the saved counter to be preserved across calls to
`rawExecute`, and is also a small step towards `finally` clauses working
as expected in the VM: raising and handling an exception or entering a
`finally` clause no longer abort clean-up in the frame above.

With both changes together, VM yields no longer interrupt unwinding or
frame cleanup.

Misc
----

- support traces with the `vmrunner` again