sem: integer literal to holey enum error detection
Previously range error detection for integer literal to holey enums
were in `semfold`, this is now done in `semConv`.

This simplifies `semfold` and moves this logic to the earliest point
when we have enough information to make the correct judgement. Added
some tests to avoid regressions.

Aside, noticed `lang_def` vs `lang_types` and that doesn't make sense,
will fix in a separate PR.