`semfold` uses `nkError` dropping  'reports'
The `semfold` module now issues `nkError` instead of reports, this
simplifies the code as everything is more expression oriented.

There remains one indirect dependency on reports, via the calls to
`msg.internalError`, though they should be easy to remove soon.

Additionally, `internalError` and `internalAssert` in the `msgs` module
now hide their reports dependency, ensuring that callers in modules no
longer incur a direct reports dependency.

Updated a callsite for `getConstExpr` in `sempass2` to detect if an
error is returned and then emit them. This might not always be
sufficient but bootstrap works, all tests pass, tools build, and docs.
The rest of the compiler is also becoming more `nkError` aware so gaps
should be temporary.