sem: report an error for calls to erroneous routines
Summary
=======

If overload resolution picked the symbol of a routine whose definition
(header or body) contains an error, the call expression was passed on,
resulting in AST containing `nkError` nodes to possibly `transf` and
`mirgen` (in the case of macro evaluation), crashing the compiler.

After overload resolution, `semOverloadedCall` now produces an
`nkError` if the callee's definition contains an error (identified by
the `TSym.ast` field containing an `nkError` node).

Details
=======

Both `semProcAux` and `semTemplateDef` are adjusted to also update the
`TSym.ast` with the `nkError` node, otherwise the detection of a routine
with an error is not possible.

There still exists logic that doesn't resolve macros and templates
through `semOverloadedCall` call. In order to help in finding these,
`nkError` nodes reaching `transf` or template expansion are now treated
as an internal error.

Misc
----

- fix a field-access defect in `semCustomPragma` that happened when
  `semOverloadedCall` returns an error