vmgen: separate dependency collection
Summary
=======

Move the gathering of dependencies and registering them in the link
table out of code generation into a separate step. While less
efficient, this greatly reduces the amount of state code generation and
serialization (`vmcompilerserdes`) need access to.

Details
=======

This change is a preparation for splitting up `TCtx` and for the
introduction of the unified backend processing. Most of the code
adjusted here will change again as part of implementing the
aforemnetioned tasks.

Access changes
--------------

- the code generator routines no longer mutate `TCtx.symToIndexTbl`
  and `TCtx.codegenInOut` (the latter is also not accessed anymore)
- `vmcompilerserdes.serialize` no longer queries nor mutates
  `TCtx.codegenInOut` and `TCtx.typeInfoCache`

Changes
------------

- add `gatherDependencies` and call it prior to all `vmgen` invocations
- manually register referenced routines in `putIntoReg`
- simplify the procedure processing in `vmbackend` -- collecting
  routines referenced from constants now happens as part of
  `gatherDependencies`, removing the need for the loop
- move `flags` from `CodegenInOut` into `TCtx`
- rename `CodegenInOut` to `LinkState`

Misc
----

- remove dead code in `vmgen.genVarSection`