sem: refactor and fix `semArrayConstr`
Summary
=======

Refactor the `semArrayConstr` procedure, make it propagate `nkError`
nodes, and fix a type inference issue: if the first element in an array
constructor was a `sink T` parameter, the array's element type was
inferred to be `sink T`, which resulted in a `type not allowed` error.

Details
=======

- move array element analysis to a separate procedure
- wrap the result in `semArrayConstr` in an `nkError`, if necessary
- improve the documentation of `semArrayConstr`
- add `semRealConstExpr`, meant for expressions that *must* be constant
- remove the `TAstDiag.maxOrdIdx` field -- the value is now taken from
  the provided `ordRange`
- don't treat a `nkConstDef` as having no init part if the init part is
  erroneous. This fixes a `constant expression expected` error always
  replacing the actual error

Known issue
-----------

The (pre-existing) logic for rejecting up- or down-conversion of
objects in an array constructor doesn't skip `sink`, alias, and
generic instance types, which is incorrect. Correctly skipping the
types uncovered that a test is depending on the previous behaviour, and
called into question whether disallowing object slicing in array
constructors is the right semantics.

For the time being, the incorrect logic is kept as-is.