language: remove new-with-finalize
Summary
=======

The new-with-finalize (`new(x, finalizer)`) feature had unintuitive
behaviour, worked different when using `--gc:arc|orc`, and was unsupported
both in compile-time execution contexts and by the JavaScript and VM
backend.

Due to the amount of issues the feature had, and destructors being able
to replace it in all valid use-cases, the feature is removed.

Details
======

- remove the `mNewFinalize` magic and adjust all usage sites
- replace finalizer usage with destructors for the `nre` and `re` module
- remove tests for new-with-finalize feature
- adjust tests that use new-with-finalize but don't depend on it