`considerQuotedIdent` is no longer call recursive
Rewrite `lookups.considerQuotedIdent` in preparation for further
structuring AST diagnostics. The new version eliminates call recursion.

Also included a variety of tests, which unfortunatley indicate a number
of gaps in the spec. In its current form the feature is occasionally
useful, but often far too optimistic.

A minor improvement to `nimsuggest` was shipped as part of this to
slightly reduce memory allocation spikes from reports.

To Revisit in the Future:
Macros can create grammatically invalid structures, including for
identifiers, where should we check for identifier well formed-ness? The
previous proc was slightly more forgiving but not by much.