vmgen: fix `set `elements with large offsets
Summary
=======

Using a non-static value as a `set` element operand to procedures like
`in`, `incl`, etc. resulted in an internal compiler when the lower
bound of the target `set` had an offset outside of the -127..127 range.

This was caused by the code generator incorrectly assuming that the
offset would fit into the immediate operand of an instruction using the
ABI encoding.

The code generator now selects the appropriate instruction for applying
the offset, considering the offset value.