custom numeric literal parsing no longer crashes
Parsing for custom numeric literals led to a compiler crash as the logic
resulted in a field access defect. That's been fixed and the tests
adjusted.

Updated `toPNode` routine to no longer attempt to access the
`isBlockArg` and `sons` fields for a parsed custom literal as those
fields do not exist on that node type.

Note, testing was finding the issue before, sort of. The reason why it
went unnoticed was because `tcustom_numeric_literals` reported a pass
because we first test it via a `static: ` call. This likely caused the
semfold/VM to not only suppress the errors but potentially modify the
AST such that runtime never executes the code. As a workaround for
this, until the real reason is found I've inverted the execution,
ensuring runtime is first.