`lang_callable` tests enabled for all backends
Summary

Enabled all tests in the `lang_callable` test category for all backends
by default. Some tests might be a platform specific here and there, but
they generally run across the board. This work is to increase test
coverage when enabling closure iterators across all targets.

Details

This resulted in a number of test modifications, such as dropping usage
of `system.stdout` for output. This was replaced with appending to
strings and/or `echo` as necessary. For disabling specific tests or
subsections comments were recorded with the keyword "knownIssue".

Replacing all `echo`ed output with `doAssert`s wasn't attempted, as the
scope of change would be too large.

Follow-Ups:
- false negatives due to VM floating point to string format differences
- certain imports json/stream/os seem cause VM target failures