support `.closure` iterators with all backends
Summary
=======

`jsgen` and `vmgen` now implement the `finished` magic procedure, making
`.closure` iterators available:
- at compile time
- for the NimScript target
- for the standalone VM target
- for the JavaScript target

Details
=======

The `finished` magic procedure is responsible for querying wether the
provided iterator has finished, and is inserted by the compiler when
transforming for-loops that use `.closure` iterators.

Implementing support for it in the JavaScript code-generator is
straightforward, as the operation can be mapped to a simple field lookup
 + comparison.

Closures (which includes `.closure` iterator instances) are not
represented as normal tuples in the VM, but rather via a dedicated type
that uses dedicated operations for querying and modification. To be able
to access the environment of a closure object, the new ``AccessEnv``
operation is introduced, which the `mFinished` implementation in `vmgen`
then makes use of.

The compiler run-time procedure `closureIterExcSetup` is also
implemented for both the JavaScript and VM target, which required the VM
callback handling slightly adjusted in order to support setting the
virtual machine's current exception.

Tests
-----

- enable the tests that make use of `.closure` iterators for the
  JavaScript and VM target
- reduce the test for a closure bug, removing the unnecessary dependency
  on the `os` module. Since the bug is unrelated to `.closure`
  iterators, it is moved to a different test file
- move a test unrelated to `hash` out of `thashes.nim`
- make a test in `ttemplate_issues.nim` work when located inside a
  `block`