sem: fix default parameter handling for macros
Summary
=======

Default parameters did not work for macros that are invoked without
overload resolution happening prior, resulting in `wrong number of
arguments` errors. This is fixed now, meaning that invoking macros for
which all parameters have default values now works even when not using
the call or command syntax.

Details
=======

`semMacroExpr` now uses the `evalTemplateArgs` procedure for pre-
processing and validating the arguments.

With the usage of `evalTemplateArgs`, the previous argument validation
logic in `semMacroExpr` and `evalMacroCall` becomes redundant and is
removed, reducing the amount of code duplication a bit.

In addition, the recursion check in `evalMacroCall` is removed. The
procedure is not re-entrant, which made the guard unnecessary. Infinite
nesting (i.e.: a macro evaluating to a call to itself) is already
detected by `semAfterMacroCall`.

Issues discovered
-----------------

Default parameters are *always* analysed during routine header analysis,
even for ``untyped`` parameters. In practice, this means that AST passed
to ``untyped`` parameters with default values can not be relied on to
really be untyped.