internal: begin moving `newSymG` to `newSymGNode`
Summary

Internal refactoring for how definition position `nkIdentKinds` are
handled in semantic analysis. Added `newSymGNode` as a replacement of
`newSymG` which returns a canonical symbol node or an error and no
longer does any legacy reporting. As part of this change also introduced
an approach that consolidates diagnostic kinds and let us do more error
recovery (for `check`, `suggest`, etc use cases).

Details

As mentioned added `newSymGNode`, `semRoutine` now uses it instead of
`newSymG`, in follow-up changes we'll be able to drop a number of legacy
reports.

Unlike `newSymG` which produces a `PSym` along with report side-effects,
`newSymGNode` only returns a `PNode`, which might be an error. What's
different about `newSymGNode` is that it only ever returns a single
diagnostic kind `adSemDefNameSym`. This singular kind describes all
possible errors as subvariants, which no longer run into all sorts of
naming colisions or giant `TAstDiag` variant of doom. With this
approach, common fields were easily setup and a singular recovery symbol
field was provided, along callees to trivially continue to make progress
and have easy access to the original input.

As part of error generation, `newSymGNode` will not modify existing
symbols and instead will create a copy of the original to avoid mutating
pre-existing symbols.

With the above change, lambda kinds which previously assumed that all
lambdas are to be treated as `skProc` triggered the error condition,
resulting in a copy being made. Copies have a "desemming" affect where
any previously attached AST is dropped. Instead of copying over the AST
on error, the `nkLambdaKinds` branch in `semexprs` was modified with a
heuristic to determine whether an `skFunc`, `skIterator` or `skProc`
should be asked for when calling `newSymG`.

Additional improvements made while working on this change:
- `semOverride`: replace node param with info
- `semOverride`: move `prevDestructor` & `bindTypeHook` helpers inside
- `semstmts.swapResult`: single use; now inner proc
- `errorreporting.localReport` add `line` pragma for better stacktrace

Follow-Up Work

In no particular order:
- update all other `newSymG` callsites
- update `considerQuotedIdent` to behave similarly
- similarily update `semIdentVis` and other definition name analyses

Longer-Term

Consolidated diagnostic kinds should allow for far more error recovery
and better encode the expected ordering of analyses/diagnostics.

Observations

`checkSpecialOperators` is `adSemDefinitionNameSym` aware. When doing
this I realized there is likely two dimensions along which we want to
organize diagnostic kinds. One is grouping based on AST position, or the
kind of analysis, the other is within that the fidelity, or timing, of
the analysis. A higher fidelity analysis should result in diagnostics
that are aware of the lower fidelidty ones in the same position. So in
this case, `checkSpecialOperators` like `newSymGNode` both operate over
'definition name' positions, so their diagnostic kinds should fall under
the same branch of the hierarchy, but within that branch, the analysis
carried out by, and resulting diagnostics, `checkSpecialOperators` must
be `newSymGNode` or another other lower fidelity analysis/diagnostic.
By having an ordering of diagnostic kinds we easily check and wrap
error nodes, which will let us compose them.