compiler: improve macro parameter handling
Summary
=======

This commit introduces internal signatures for macros, removing
conditional logic from parameter symbol handling and generally moving
the details of how macro parameters work out of `vmgen` and more into a
single place in `sem`.

As two side-effect of the above:
- inspecting parameter types from a macro's header (via, for example,
  `typeof`) now yields the types as specified, instead of as `NimNode`
- the types (or values, in case of `static[T]`) passed to a macro's
  generic parameters can now be inspected by the macro

Details
=======

The symbols belonging to macro parameters were previously replaced with
new ones that used either `NimNode` or, in the case of `static[T]`, the
base type as the type, prior to adding them to the symbol table.

Replacing the symbols at some point is necessary in order for semantic
analysis inside the macro body to view them as the correct type, but the
previous implementation had two issues:
1. the parameters were only visible as `NimNode` to other types or
   expressions referencing them
2. the symbols were freestanding. They weren't registered anywhere,
   meaning that it was not possible to, for example, map all parameter
   symbols via their ID to some other entity

To fix both issues and in order for the implementation to be less
scattered, a different approach is now used. Macro parameter handling is
moved out of routine header analysis into `semMacroDef` (fixing #1).
This allows for a more flexible implementation (more contextual
information is available) and also reduces the amount of context
required by `addParam`, etc.

So that they can be later queried, the created internal symbols are
added as parameters to a dedicated `tyProc` type that represents the
macro's internal signature, which is then stored as part of the macro's
symbol.

Changes
-------

- remove macro parameter handling from `addParamOrResult` and
  `addResult`
- add a dedicated branch for `skMacro` to `TSym` and adjust the places
  depending on the previous layout
- create the internal symbols for macro parameters in `semMacroDef`,
  where they're also used to populate the internal signature
- add a `transformBody` variant that doesn't query nor updates the
  symbols `transformBody` field
- change `vmgen` to consider the macros internal signature, making the
  previously used workarounds obsolete
- rewrite `getAst(macro(a, b, c))` during `mirgen`, slightly simplifying
  the related logic in `vmgen`
- remove the now unused `adVmGenBadExpandToAstCallExprRequired`
  diagnostic