vm: improve and fix numeric conversion support
Summary
=======

Rework the handling of conversion between two numeric values (integers,
floats, enums, etc.) in the VM and `vmgen`, fixing two severe issues:
- unsigned integers greater than 2^63 being treated as signed values
  during unsigned-to-float conversions
- range checks being erroneously inserted for X-to-unsigned conversions

Numeric conversions now work the same for the VM target as they do with
the C target. In addition, they no longer use the high-quality RTTI,
thus making them more efficient.

Details
=======

One problem with the previous `Conv` operation was, that the semantics
regarding the destination register were dependent on the target type.
For to-`string` and to-`openArray` conversions, the destination register
was expected to store a handle, while for the numeric conversions it was
an "out" register (meaning that its content is overwritten).

Because the operation unconditionally validated the handle stored in
destination (even it was only meant as an "out" register), this would
lead to spurious access violation errors if the destination register
stores a stale handle.

In addition, numeric conversions didn't properly clean up the
destination register, which was a cause of memory leaks -- these are
fixed now.

- introduce the new `NumConv` operation, which is used for conversion
  involving floats, and for the to-bool conversion
- adjust numerical conversion handling in `vmgen` to emit `NumConv`
  where applicable
- remove the numeric-conversion-related logic from `opConv`
- change the `Conv` and `Cast` operation to use two instruction words
  instead of three
- only emit `RangeCheck` instructions for `nkChckRange` expression if
  the destination is not an unsigned integer, mirroring the C code-
  generator's logic
- treat `range`-to-base-type conversions as a no-op in `vmgen`
- split the `cast` generation logic into a separate procedure
- unsupported `cast`s are now reported during code generation instead of
  at run-time