transf: de-duplicate non-routine symbols
Summary
=======

Rewrite var, let, and const definitions to use a copied symbol if the
symbol already appeared in a definition position once, and redirect
following usages of the original symbol to said copies.

Until semantic analysis properly sanitizes the AST after template or
macro expansion, this works around a common internal compiler error, and
is a preparation for switching to ORC as the default memory management
strategy.

Details
=======

Substituting template parameters with typed AST effectively leads to
symbol redefinition (although it is not reported as an error).

This causes issues for the `injectdestructors` pass and the code
generators, as both operate under the assumption that a symbol only
appears exactly once in a definition context.

`transf` is chosen for fixing-up the symbols because:
- it has access to the owning module's `IdGenerator` (`mirgen` doesn't)
- it already has a facility for rewriting symbol nodes in usage
  positions

Only globals, constants, and locals not processed by lambda-lifting are
de-duplicated however -- other redefined symbols (like those of
routines) will still cause issues.