stdlib: fix `range` types for `repr_v2`
Summary
=======

Fix an infinite template expansion error when passing a range type with
a `distinct T` base type to `repr` when using `repr_v2`.

In order to support to fix the `repr` the implementation, the new
type-trait routine `rangeBase` is introduced. As the name implies, it
used for querying a `range` type's base type.

Details
=======

For an expression of type `range T` where `T` is a `distinct`, the
template `repr(x: distinct)` is chosen by overload resolution.

But since `range T` is not a `distinct` type, `distinctBase` returns the
same type as is passed in, which causes the type to be left unchanged
and the same `repr` overload to be selected again.