vm: improve support for `var`, `lent`, and `addr`
Summary
=======

- support implicit and explicit to-`pointer` conversions
- the `ptr` value produced by taking the address of a primitive local
  (e.g. int, float, etc.) can now be assigned to a field or array-slot.
  An internal error was previously raised by the VM at run-time
- fix many issues in edge-cases related to `var`|`lent` parameters,
  return values, and views
- fix pointers to primitive locals not being checked for memory access
  violations (e.g. using a pointer to access a local that went out of
  scope wasn't detected as a violation)
- fix an evaluation-order issue with parseBiggestFloat in the VM
- for-vars (i.e. the i in for i in ...) are now properly treated as
  globals
- significantly improve efficiency of accessing `var T` parameters where
  `T` is a compound type or string
- improve the efficiency of assigning to global variables

Note that all of the above fixes and improvements only apply to code
running in the VM -- the C and JS targets are not affected.

Details
=======

Motivation
----------

The motiviation behind the rework is to get rid of the very costly
handle <-> address overhead, which was previously incurred on every
`var T` parameter access where `T` is a non-primitive type.

Another goal is the removal of the register-addresses (`rkRegAddr`),
which, aside from complicating the VM's design, have multiple issues:
- they're not subjected to the standard access checks
- they require registers to be stored at fixed host memory locations,
  which prevents various optimizations to the VM
- they make pointers in the VM dynamically typed: a pointer can be
  either a "real" address or a register-address

`vmgen` changes
---------------

- remove the remaining `gfNode` flag. It is replaced with either a
  `load` parameter or with using separate procedures
- don't unconditionally treat `var` and `lent` types as pointers
  - `var T` parameters and return variables use handles
  - fields and globals that are direct views are stored as pointers
- code generation for view assignments and `nkHiddenDeref` takes the
  above difference into account
- all locals that need to be addressable via a handle or pointer (they
  have their address taken or are passed to a ``var`` parameter) are
  prevented from using register storage. This only applies to locals
  of primitive type (ints, floats, `NimNode`, and pointers), as the
  others never use register storage
- since its now obsolete, the logic for "patching assignments" is
  removed
- magics that mutate their operands but don't use `nkHiddenAddr`
  require special "write back" logic in order to support their
  operands using stack-cell storage. This is implemented via `Loc`
  and `genLoc`

The approach to supporting `addr` and views for locals of primitive
type is similar to the one employed by `jsgen`, but instead of relying
on the `sfAddrTaken` flag, an AST-based analysis for gathering all
locals that need to be addressable via a handle or pointer is run prior
to code generation.

Result variable and immutable parameters of primitive type that need to
be addressable via handles or pointers require dedicated handling:
- the result variable uses stack-cell storage and is transitioned to
  register storage prior to non-exceptional procedure exits
- immutable parameter registers are transitioned to stack-cell storage
  in the procedure preamble

VM changes
----------

- replace `opcAddrReg` and `opcAddrNode` with the new, simpler `Addr`
  operation
- remove the `rkRegAddr` register kind and all workarounds related to
  it
- remove the `LdGlobalAddr` operation -- it is replaced with usage of
  `LdGlobal` + `Addr`

Views vs. `var` parameters
--------------------------

The way both are now implemented by `vmgen` indicates that the notion
started by the MIR of them being two different things is the right
direction.

The difference is not fully enforced through the AST reaching the code
generator yet, which leads to much of the handling being implicit,
making the implementation harder to follow.

This will eventually be improved upon by either basing the code
generator on the MIR or (more likely) a dedicated IR.