ast: introduce the `CheckedAst` type
Summary
=======

It is intended to make error propagation and handling easier and less
manual. This commit only adds the type and some basic routines for
interacting with it -- no other code is adjusted to make use of it.

So far, the API only covers the very basics, and should be extended on
an as-needed basis.

Details
=======

The plan is to incrementally adjust all semantic analysis procedures to
return a `CheckedAst` instance, which should eventually allow for the
removal of generic wrapper errors (i.e. `adWrappedError`).

In addition, it could also become useful for storing warnings, or
for encapsulating other side-effects produced by sem-checking a piece of
AST.