sem: disallow `lent|sink T` in non-type contexts
Summary
=======

A `lent|sink T` expression is represented as a
`Prefix (Ident ...) (...)` in untyped AST, and was analyzed in
`semConv`.

Besides this being unintuitive, inefficient, and also inconsistent with
other modifiers like `var`, it also allowed `lent|sink T` type
expressions in non-type contexts.

Instead, `lent T` and `sink T` type expression are now analyzed as part
of `semTypeNode`, which makes their handling more consistent with that
of `var T`.

Details
=======

The `lent2` helper template in the `iterators_1.nim` module is adjusted
to use `typedesc` templates. This is now required because `lent T` can
no longer be used in a non-type context (which currently includes
`untyped` templates being expanded in type positions).