sem: propagate `nkError` in `semCast`
Summary
=======

- make `semCast` propagate error nodes
- introduce `semTypeNode2`, which is similar to `semTypeNode`, but
  returns a node instead of a type
- handle error types in `isCastable`, `productReachable`, and
  `iterOverTypeAux`

AST types
---------

- extend the `CheckedAst` API with `typ` accessors and more ways to
  initialize
- disallow and prevent `CheckedAst` from storing an error node directly
- add the experimental ``ElaborateAst`` type (name is going to change)

Details
=======

The idea with `semTypeNode2` is that it should eventually replace
`semTypeNode`, with all of `semtypes` being refactored to return nodes
instead of types. This would allow the type-node usage sites to store
typed AST, would reduce the amount of input AST modifications, and make
error propagation in type node contexts a bit simpler.

A lesson learned from the `tsequtils.nim` test failure is that error
types (`tyError`) are sometimes either dropped or reported too late.
The direction that this commit goes into in order to prevent this, is
to produce an error node for all erroneous type-expressions appearing
in the AST.